Refactor module loading, make proper package with deps
