+++ .gitignore
@@ -1,0 +1,23 @@
+*.DS_Store
+*.jl.cov
+*.jl.*.cov
+*.jl.mem
+*.DS_Store
+deps/deps.jl
+docs/build/
+docs/site/
+src/**/Manifest.toml
+*.ipynb_checkpoints/
+*.vtk
+*.dat
+*.vtu
+*.pvtu
+*.swp
+*.jpg
+*.png
+*.nc
+!pics/*.png
+!pics/*.jpg
+!docs/src/assets/*.png
+*~
+TAGS
