+++ test/runtests.jl
@@ -1,0 +1,10 @@
+using Test
+using Land
+
+# include("testODE.jl")
+# include("testODE_v2.jl")
+
+@testset "Land" begin
+    @test 1==1
+end
+
